ai-accounting-agent/pyproject.toml
[project]
name = "ai-accounting-agent"
version = "0.1.0"
description = "Build an AI agent that completes accounting tasks in Tripletex. You receive a task prompt (in one of 7 languages), use the Tripletex API to execute it, and get scored on correctness and efficiency."
readme = "README.md"
authors = [
    { name = "valiantlynx", email = "gormerykombo@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "ai-accountant-mcp",
    "lancedb>=0.30.1",
    "pydantic-ai>=1.70.0",
    "google-genai>=1.68.0",
]

[tool.uv.sources]
ai-accountant-mcp = { path = "ai-accountant-mcp" }

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
